% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Em_f1}
\alias{Em_f1}
\title{Use EM algorithm to estimate the parameters A and Pi of an HMM}
\usage{
Em_f1(m, A, Pi, f0x, f1x, fw_bc_EM, x, eps, maxit, h)
}
\arguments{
\item{m}{the number of positions (hypothesis)}

\item{A}{a matrix 2 * 2 the transition probabilities}

\item{Pi}{a vector of the initial state probabilities}

\item{f0x}{a vector of the values of the density under the null hypothesis on the observations}

\item{f1x}{a vector of the values of the density under the alternative hypothesis on the observations}

\item{eps}{the value ta reach for the convergence}

\item{maxit}{integer, the maximum number of iteration}

\item{alpha}{a matrix m * 2  containing the forward variables}

\item{beta}{a matrix m * 2  containing the backward variables}
}
\value{
Product of matrices
}
\description{
Use EM algorithm to estimate the parameters A and Pi of an HMM
}
\examples{
 m <-  10
 A <- matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T)
 f0 <- c(0, 1)
 f1 <- c(2, 1)
 Pi <- c( 0.9, 0.1)
 rdata <- simulate.data.hmm.2states(m, Pi, A, f0, f1)
 x <- rdata$x
 theta <- rdata$theta
 mod <- for_back(m, A, f0x, f1x, Pi)
 f0x <- dnorm(x, f0[1], f0[2])
 f1x <- dnorm(x, f1[1], f1[2])
 alpha <- mod$alpha
 beta <- mod$beta
}
