% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_L1}
\alias{get_L1}
\title{Simple matrix multiplication}
\usage{
get_L1(A, m, alpha, beta, f0x, f1x)
}
\arguments{
\item{A}{a matrix 2 * 2 the transition probabilities}

\item{m}{the number of positions (hypothesis)}

\item{alpha}{a matrix m * 2  containing the forward variables}

\item{beta}{a matrix m * 2  containing the backward variables}

\item{f0x}{a vector of the values of the density under the null hypothesis on the observations}

\item{f1x}{a vector of the values of the density under the alternative hypothesis on the observations}
}
\value{
Product of matrices
}
\description{
Simple matrix multiplication
}
\examples{
A <- matrix(1:9, 3, 3)
B <- matrix(11:19, 3, 3)
matrix_mult_cpp(A, B)
}
