% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getbound}
\alias{getbound}
\title{New way of finding Bin ! (Now A)}
\usage{
getbound(m, alpha, li0, f0x, f1x, Pis)
}
\arguments{
\item{m}{the number of positions (hypothesis)}

\item{alpha}{a matrix m * 2  containing the forward variables}

\item{f0x}{a vector of the values of the density under the null hypothesis on the observations}

\item{f1x}{a vector of the values of the density under the alternative hypothesis on the observations}

\item{A}{a matrix 2 * 2 the transition probabilities}

\item{beta}{a matrix m * 2  containing the backward variables}
}
\value{
Product of matrices
}
\description{
New way of finding Bin ! (Now A)
}
\examples{
A <- matrix(1:9, 3, 3)
B <- matrix(11:19, 3, 3)
matrix_mult_cpp(A, B)
}
