% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Selection_delta}
\alias{Selection_delta}
\title{Title}
\usage{
Selection_delta(
  x,
  fw_bc,
  seuil,
  A_est,
  f0x_est,
  f1x_est,
  Pi_est,
  min_size,
  min_jump = NULL,
  pval = NULL,
  all = FALSE
)
}
\arguments{
\item{x}{vector of statistics}

\item{fw_bc}{result of}

\item{seuil}{threshold for p values}

\item{A_est}{the estimation of the transition matrix}

\item{f0x_est}{estimation of the null density}

\item{f1x_est}{estimation of the alternative density}

\item{Pi_est}{estimation of the initial distribution}

\item{min_size}{keeping only a given number of}

\item{min_jump}{}

\item{pval}{a vector of pvalue}

\item{all}{does the set selecting all the hypothesis have to be add ?}
}
\value{

}
\description{
Title
}
