% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utile.R
\name{sim_markov}
\alias{sim_markov}
\title{Simple matrix multiplication}
\usage{
sim_markov(m, Pi, A)

sim_markov(m, Pi, A)
}
\arguments{
\item{m}{size}

\item{Pi}{initial distribution}

\item{A}{transition matrix}

\item{B}{Matrix}
}
\value{
Product of matrices


}
\description{
Simple matrix multiplication

Title
}
\examples{
A <- matrix(1:9, 3, 3)
B <- matrix(11:19, 3, 3)
matrix_mult_cpp(A, B)
}
